Adding code to compute the next schema given the existing and inserted schema.
Right now I have a main function used for testing and I am taking in
existing and inserted schemas in as string form. I generate the output
schema in a vector containing both of the column names and other
relevant information such as the type, m, d and if it is unsigned.