Adding dummy merge changes for adding columns.
Right now, the adding columns does not work. It SIGABRTs when I try and
execute the alter table statement when trying to update to a new schema
from an existing schema. It seems to not allow me to run the alter
statement when I am doing the changes and adding columns and then
adding the dummy columns. I'm currently looking into this.