Works when adding or modifying types that do not fit in the existing dummy columns. Does not work for renaming the dummy columns and then dropping and creating a new view that includes these.
For some reason it is not adding any of the inserted data to these
columns. I do get an error after running my code and I believe it
happens when trying to load the data into a renamed column. The error
says: ERROR 1611 (HY000): Invalid column reference
(adults2.relationship) in LOAD DATA. I'm not sure why this is happening
as the view correctly has these columns with the correct types.